refactor model/
