+++ src/Effects.ts
@@ -1,6 +1,6 @@
 // eslint-disable-next-line import/no-cycle
 import { StoreEffects } from './Store';
-import { IUserRepository } from './domain/models/User/IUserRepository';
+import { IUserRepository } from './domain/repository/user/IUserRepository';
 import UserRepository from './infrastructure/repository/UserRepositoryImpl';
 
 const userRepository: IUserRepository = new UserRepository();
